refactor: utils structure
